Version 1 (not yet run)
This is a progress commit. The code is mostly done but I was not able to run this yet due to my OS not allowing me to install the ProgLearn package. I've now been granted access to the synaptomes cluster, so that I can run the notebook there. So, for next week, I'll run and bug fix it using this cluster (rather than on my machine).