Bayes' Error Estimation for Gaussian Parity
Added Jupyter Notebook which traces the Bayes' error estimate for the Gaussian Parity problem. 
Analytical steps are mentioned, and the final estimation is done by a numeric estimation.

This is a standalone notebook, and can be used for different projects as it focuses on calculation of Bayes' Error for the Gaussian XOR problem only.